# Free typed arrays for instanced models