--- a/Source/Scene/Model/NodeStatisticsPipelineStage.js
+++ b/Source/Scene/Model/NodeStatisticsPipelineStage.js
@@ -1,12 +1,12 @@
 import defined from "../../Core/defined.js";
 
 /**
- * The node statistics update stage updates memory usage statistics for
- * Model on the node level. This counts the binary resources
- * that exist for the lifetime of the Model (e.g. attributes
- * loaded by GltfLoader). It does not count resources that are created
- * every time the pipeline is run. The individual pipeline stages are
- * responsible for keeping track of additional memory they allocate.
+ * The node statistics update stage updates memory usage statistics for a Model
+ * on the node level. This counts the binary resources that exist for the
+ * lifetime of the Model (e.g. attributes loaded by GltfLoader). It does not
+ * count resources that are created every time the pipeline is run.
+ * The individual pipeline stages are responsible for keeping track of any
+ * additional memory they allocate.
  *
  * @namespace NodeStatisticsPipelineStage
  *
@@ -20,15 +20,15 @@ NodeStatisticsPipelineStage.process = function (
   node,
   frameState
 ) {
-  const model = renderResources.model;
-  const statistics = model.statistics;
+  const statistics = renderResources.model.statistics;
+  const instances = node.instances;
   const runtimeNode = renderResources.runtimeNode;
 
-  countInstancingAttributes(statistics, node.instances);
+  countInstancingAttributes(statistics, instances, runtimeNode);
   countInstancing2DBuffers(statistics, runtimeNode);
 };
 
-function countInstancingAttributes(statistics, instances) {
+function countInstancingAttributes(statistics, instances, runtimeNode) {
   if (!defined(instances)) {
     return;
   }
@@ -38,24 +38,31 @@ function countInstancingAttributes(statistics, instances) {
   for (let i = 0; i < length; i++) {
     const attribute = attributes[i];
     if (defined(attribute.buffer)) {
-      // Packed typed arrays are not counted
+      // Any typed arrays should have been unloaded before this stage.
       const hasCpuCopy = false;
       statistics.addBuffer(attribute.buffer, hasCpuCopy);
     }
+  }
+
+  if (defined(runtimeNode.instancingTransformsBuffer)) {
+    // The typed array containing the computed transforms isn't saved
+    // after the buffer is created.
+    const hasCpuCopy = false;
+    statistics.addBuffer(runtimeNode.instancingTransformsBuffer, hasCpuCopy);
   }
 }
 
 function countInstancing2DBuffers(statistics, runtimeNode) {
   if (defined(runtimeNode.instancingTransformsBuffer2D)) {
-    // The typed array containing the computed 2D transforms
-    // isn't saved after the buffer is created.
+    // The typed array containing the computed 2D transforms isn't saved
+    // after the buffer is created.
     const hasCpuCopy = false;
     statistics.addBuffer(runtimeNode.instancingTransformsBuffer2D, hasCpuCopy);
   }
 
   if (defined(runtimeNode.instancingTranslationBuffer2D)) {
-    // The typed array containing the computed 2D translations
-    // isn't saved after the buffer is created.
+    // The typed array containing the computed 2D translations isn't saved
+    // after the buffer is created.
     const hasCpuCopy = false;
     statistics.addBuffer(runtimeNode.instancingTranslationBuffer2D, hasCpuCopy);
   }