Closes #10421. This addresses the temporary buffers created by `InstancingPipelineStage`. The attributes are loaded as typed arrays for reasons like computing transform matrices, or computing the min/max of the translation attribute if it's not provided. But afterwards, there is no further manipulation of the data, and thus no reason to keep recreating the buffers from those typed arrays.

This PR rewrites parts of `I3dmLoader` and `GltfLoader` to load in transform attributes as typed arrays _only if_ the rotation attribute is defined (i.e. matrices have to be computed). Additionally, feature IDs were for some reason being loaded as typed arrays, so they have been restricted to loading as buffers instead.

**TODO**: need to profile the changes for `i3dm`s in combination with other functions (those that require rebuilding draw commands)