@j9liu 2022-08-29
Did some performance testing. Here's the difference between heap screenshots, using this [Sandcastle](http://localhost:8080/Apps/Sandcastle/index.html#c=jVZtc9o4EP4runwyM1T4DdukSecS8lIzNUkD6UuGmRthC1CRJZ8sk5Cb/PeTZENxLr3pB8Del2d3n12tSDkrJdgQ/IgFOAUMP4IhLkmVwy9GZs2OUvM+5EwiwrCYHXXBPzMGwJLyOT4GC0RLPGMvnfczNmO9Xo0F8ZPELLMasFpYv3gXU0JxGbOywKnkIiFPhGnvXi812ZCdRqeQ8AxTlVkDm74JYfCNZZMEuBaIyaaU0QSgVDmWQHKw5ZUAhDOAyhLLcsaaDGPOYIYXqKLyzBhP+RozFXh2hLej1fw6JTdkFN8/x86YxCr0XT8dxkG8Lr59GY4GUBn9nV2vlVH8PM6vVsnzQ55MHJL8WG8/TT/7yUUib75ePt8MnfVDfvUj+fr96SG/346/xuTTcFQ8KLDxxecyzukqU8/J9PvjeHrmJNOz/tizYRZ+y8fzqzi/vfTO+2Q5+ja6Dx+iC3Ruh8V5+eNuvnY+Sjm9qNazI8PCjC0qlkpdK+UosyiaY9oFlaCduoE121IziWW7+a1OYWkZe6B99YPpNQC9XuMLc/RE8iqfpAJjNilQii+F4HqgfGPZNK9MMcOwECQnkmxwCVGWWQ2Ghnypm2eyVbRTvEE6fT1xP7t0h0vVwhTDheD5mW5inFlOFPQHdqeeosZ/WZqJKH/LPfQi/5W7rN3PBcmW+LdAPDeoQRoIqegof9ez36lP0G7QUY4Fggu6nfKa/gyXkjDDCDhuNQsJqZ4Q86x3rt93B4EDg8D1HN/37aAL3vlhOHAcB3qB7w+8sO93gee76teBoeu7gTPwoqDT1UG4IFgd9CZI03YVmwic/m9kG/r2wOk7vu14A98ZuLaK8k5JHb8fBYHj+V7oqO++kUaOG/muVgR+OPCjOrgZseLXxdnQjWwvcMPAC/qqZYFTxxjYfuQrIC8MXTcKoi6woefaXjQIPTuKPM9zmwgv5jurxK5C+2BzgZgtuMhrFZrzSh2OFQZpJYTihG7BiixXVH0kzsACI1kJtfjqU3SoOq15ayyOQcXUYlGrM9sxvFRtpENOuWiXqiWWTvRll5ERAbTDAltMKX8EKr8V36gp2Y9LeXD0Nirbj4hlVMvVgLGiUjtNV2XtNwJnCa9KnCgUS2eVq4dc+c1Ysxw0HYuDyI+oBIXAG6Lc2lR0dYHcULWXagSyALv139RvHXjBBrjT2U3ZG0qYmvpPW7oWge9/6au89sQbK7NdTGG3JF2r0jT1+zbu1mGhdDi72mO0V5dSWjuqoLrfbnlJNJ/1QtQV//Gq5Bbez2IFVu/sVV4fd1UYMg9Se7u8FrSBOhwkmFLOcDt+zWf313TWdbzho8K1wM36osUKWS3x3eWFPnx9s827+8t18mo6p9sCw+TmfnL5V3Lz5bLz/qh7dFLKLcUfdnvgT5IXXEh95VgQ9iTOC4rUJujNK5WchGlZdprmn/QOXU8ysgEkO33jjwtIqbr1lWZRUTohz2qrfzjpKfv/uOoFTtjyRp0xirbabOV8+FQLIYQnPfX6tqfknM6ReIX8Lw):

| before changes (main) | after changes |
| ------------------------- | -------------- |
| ![arraybuffer before changes](https://user-images.githubusercontent.com/32226860/187246264-a034eac5-c33a-486b-b3e4-c6e02a2ba0bf.png) | ![arraybuffer after changes](https://user-images.githubusercontent.com/32226860/187246272-aa57e124-3c15-4fe5-b32f-c092e79d3c66.png) |
| 57 array buffers | 46 array buffers |

For runtime performance, I used this [Sandcastle](http://localhost:8080/Apps/Sandcastle/index.html#c=jVVbU9s4FP4rmjyFmSBblnxrKbMQLusMCS0JhWbyothKIiJLWUsOJDv895UvdAvLdniRpaPzfefo3JwqqQ3YcvbICvAFSPYI+kzzMoffa1l31knrc19JQ7lkxazTA3/PJABLoebsE1hQodlMPh98nsmZdJyGC7Inw2TWbckaYXPAZxMumE6k3rDUqGLIn7is0I6T1t7wl5vKhaHKmLCetbTpuxQ1f63ZOgEuCypN+5TBGNDUAjUwCuxUWQCuJKBaM6NnsvUwURJmbEFLYU5q5YlaM2kNzzpsN1jNL1N+zQfJ7T5BI55Y0zd+2k+CZL25/94fxNAq/ZVdrq1Ssh/lF6vhfpoPx4gPH9a7q8k3Mjwbmuu78/11H62n+cXD8O7H0zS/3Y3uEn7VH2ymlmx09k0nuVhldj+c/HgcTU7QcHLij7ALs/A+H80vkvzrOT71+XJwP7gNp9EZPXXDzal+uJmv0Z/GTM7K9axTR0EwA0wVJmbseVHK1FTvFopmXUHnTPRAWYiDJpmt4usSeJUvZrq1JqhQ1abOOACO02JhTp94XubjtGBMjjc0ZedFoaqyIrVmm0KdMsngpuA5N3zLNKRZ1m05KsrnJoW1nzb4gm1p5XhVd//m6oZpm8iUwUWh8pMqlUnWRVHgx+5BU0stfqnrutAfgoc4Im/gpoGfFjxbsg+RYC9oSFoKY8OhP4r0D5o+eil3mrOCwoXYTVQT/oxpw2UdEfDpVbJoYeyOStw99IjvxQGCQeBhRAhxgx44JGEYI4QgDgiJceiTHsDEs18EQ494AYpxFBz0KiOq4My2e2ukTbu1zQuW/tayC4kbIx8RF+GYoNhzrZVDK0XEj4IAYYJDZFe/lkbIi4hXXQQkjEnUGK9LbPP/j3OhF7k48MIAB75NWYAaG7FLImKJcBh6XhREPeBC7Lk4ikPsRhHG2GstPNdrVhYvL3R/zq9mAqVKqDfj8BcPEHQr7mapYGMqs5RqI1hVyxO1XAp2WhqjpE1/c2woqyqo56X9tP3YTVcsXbOs7UO+eCP42ZpQ8OXK9FvParq6q56BLU/2W+VS2slmZ3fWAOrHdnqdI212gh2/xPwPnm9UYar27kLoGJZvBLVvdual9cfAVOum4wE4cn6FHmV8C3j25Z1fBUiFnbP2ZlEKMeZ720HHR47V/w+0ahYul9dbVgi6q9RW6PiqEUIIjxx7fB9plBJzWrxh/gc), which toggles the `lightColor` of the tileset. This forces the models to `rebuildDrawCommands`, so hopefully the difference as a result of changing `InstancingPipelineStage` would be noticeable. I recorded for 11 seconds, toggling the button 10 times within that timeframe, and got these results:

| before changes (main) | after changes |
| ------------------------- | -------------- |
| ![image](https://user-images.githubusercontent.com/32226860/187247068-48721cf8-178f-4c2e-a4ed-31cf7495222a.png) | ![image](https://user-images.githubusercontent.com/32226860/187246994-52de8299-a0e8-4539-82be-7af5dda83c9b.png) |

Overall, I believe this PR improved both memory usage and runtime performance. It's now ready for review @ptrgags